requirements.txt
mysql-connector-python==9.5.0
paho-mqtt==2.1.0
pip==25.3
